Move the HELLO tests into their own file
Eight describes covering HELLO broadcast, the round line, the three settings
riders that propagate on it, and the two initiation paths. Bodies are
unchanged; the file gets the shared preamble and the describes keep their
section banners.
